refactor(optimizer): restructure and rename nvidia-nat-config-optimizer
- Restructure package layout and flatten structure
- Rename package to nvidia-nat-config-optimizer
- Rename import path from nat.optimizer to nat.config_optimizer
- Fix doc paths and migration guide
