+++ docs/source/improve-workflows/optimizer.md
@@ -20,7 +20,7 @@ This document provides a comprehensive overview of how to use the NeMo Agent Too
 
 ## Prerequisites
 
-The optimizer is included when you install `nvidia-nat`. If you installed only `nvidia-nat-core`, install the optimizer to use `nat optimize`: run `pip install nvidia-nat-optimizer` or `pip install nvidia-nat`. See the [Install Guide](../get-started/installation.md) for details.
+The optimizer is included when you install `nvidia-nat`. If you installed only `nvidia-nat-core`, install the optimizer to use `nat optimize`: run `pip install nvidia-nat-config-optimizer` or `pip install nvidia-nat`. See the [Install Guide](../get-started/installation.md) for details.
 
 ## Introduction
