+++ examples/evaluation_and_profiling/email_phishing_analyzer/tests/test_email_phishing_analyzer.py
@@ -49,7 +49,7 @@ async def test_run_full_workflow():
 async def test_optimize_full_workflow(capsys):
     from nat.data_models.config import Config
     from nat.data_models.optimizer import OptimizerRunConfig
-    from nat.optimizer.optimizer_runtime import optimize_config
+    from nat.config_optimizer.optimizer_runtime import optimize_config
     from nat_email_phishing_analyzer.register import EmailPhishingAnalyzerConfig
 
     config_file: Path = locate_example_config(EmailPhishingAnalyzerConfig, "config_optimizer.yml")
