+++ packages/nvidia_nat_config_optimizer/pyproject.toml
@@ -27,10 +27,10 @@ git_describe_command = "git describe --long --first-parent"
 root = "../.."
 
 [project]
-name = "nvidia-nat-optimizer"
+name = "nvidia-nat-config-optimizer"
 dynamic = ["version", "dependencies", "optional-dependencies"]
 requires-python = ">=3.11,<3.14"
-description = "Prompt and parameter optimizer for NVIDIA NeMo Agent Toolkit"
+description = "Workflow config and prompt optimizer for NVIDIA NeMo Agent Toolkit"
 readme = "README.md"
 license = { text = "Apache-2.0" }
 keywords = ["ai", "rag", "agents", "optimization"]
