+++ packages/nvidia_nat_config_optimizer/src/nat/config_optimizer/optimizer_runtime.py
@@ -20,8 +20,8 @@ from pydantic import BaseModel
 from nat.cli.type_registry import GlobalTypeRegistry
 from nat.data_models.optimizer import OptimizerRunConfig
 from nat.experimental.decorators.experimental_warning_decorator import experimental
-from nat.optimizer import register  # noqa: F401 - trigger optimizer registration
-from nat.optimizer.optimizable_utils import walk_optimizables
+from nat.config_optimizer import register  # noqa: F401 - trigger optimizer registration
+from nat.config_optimizer.optimizable_utils import walk_optimizables
 from nat.runtime.loader import load_config
 
 logger = logging.getLogger(__name__)
