+++ packages/nvidia_nat_config_optimizer/src/nat/config_optimizer/prompts/base.py
@@ -1,0 +1,41 @@
+# SPDX-FileCopyrightText: Copyright (c) 2025-2026, NVIDIA CORPORATION & AFFILIATES.
+# All rights reserved.
+# SPDX-License-Identifier: Apache-2.0
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+# http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+"""ABC for prompt optimizers."""
+
+from abc import ABC
+from abc import abstractmethod
+
+from nat.data_models.config import Config
+from nat.data_models.optimizable import SearchSpace
+from nat.data_models.optimizer import OptimizerConfig
+from nat.data_models.optimizer import OptimizerRunConfig
+
+
+class BasePromptOptimizer(ABC):
+    """Interface that all prompt optimization strategies must implement."""
+
+    @abstractmethod
+    async def run(
+        self,
+        *,
+        base_cfg: Config,
+        full_space: dict[str, SearchSpace],
+        optimizer_config: OptimizerConfig,
+        opt_run_config: OptimizerRunConfig,
+    ) -> None:
+        """Run prompt optimization."""
+        ...
