+++ packages/nvidia_nat_config_optimizer/tests/test_parameter_optimizer.py
@@ -23,7 +23,7 @@ from nat.data_models.optimizable import SearchSpace
 from nat.data_models.optimizer import OptimizerConfig
 from nat.data_models.optimizer import OptimizerMetric
 from nat.data_models.optimizer import OptimizerRunConfig
-from nat.optimizer.parameter_optimizer import optimize_parameters
+from nat.config_optimizer.parameters.optimizer import optimize_parameters
 
 
 class _FakeTrial:
@@ -190,14 +190,14 @@ def test_optimize_parameters_happy_path(tmp_path: Path):
                 ("Latency", SimpleNamespace(average_score=0.5)),
             ])
 
-    with patch("nat.optimizer.parameter_optimizer.apply_suggestions",
+    with patch("nat.config_optimizer.parameters.optimizer.apply_suggestions",
                side_effect=fake_apply_suggestions) as apply_mock, \
-         patch("nat.optimizer.parameter_optimizer.pick_trial",
+         patch("nat.config_optimizer.parameters.optimizer.pick_trial",
                return_value=SimpleNamespace(params=best_params)) as pick_mock, \
-         patch("nat.optimizer.pareto_visualizer.create_pareto_visualization") as viz_mock, \
-         patch("nat.optimizer.parameter_optimizer.optuna.create_study",
+         patch("nat.config_optimizer.parameters.pareto_visualizer.create_pareto_visualization") as viz_mock, \
+         patch("nat.config_optimizer.parameters.optimizer.optuna.create_study",
                side_effect=fake_create_study) as study_mock, \
-         patch("nat.optimizer.parameter_optimizer.load_evaluation_run",
+         patch("nat.config_optimizer.parameters.optimizer.load_evaluation_run",
                return_value=_DummyEvalRun) as eval_run_mock:
 
         tuned, returned_best_params, n_trials = optimize_parameters(
@@ -299,15 +299,15 @@ class TestSamplerSelection:
                     ("Latency", SimpleNamespace(average_score=0.5)),
                 ])
 
-        with patch("nat.optimizer.parameter_optimizer.optuna.create_study",
+        with patch("nat.config_optimizer.parameters.optimizer.optuna.create_study",
                    side_effect=capture_sampler), \
-             patch("nat.optimizer.parameter_optimizer.load_evaluation_run",
+             patch("nat.config_optimizer.parameters.optimizer.load_evaluation_run",
                    return_value=_DummyEvalRun), \
-             patch("nat.optimizer.parameter_optimizer.apply_suggestions",
+             patch("nat.config_optimizer.parameters.optimizer.apply_suggestions",
                    return_value=base_cfg), \
-             patch("nat.optimizer.parameter_optimizer.pick_trial",
+             patch("nat.config_optimizer.parameters.optimizer.pick_trial",
                    return_value=SimpleNamespace(params={})), \
-             patch("nat.optimizer.pareto_visualizer.create_pareto_visualization"):
+             patch("nat.config_optimizer.parameters.pareto_visualizer.create_pareto_visualization"):
 
             optimize_parameters(base_cfg=base_cfg,
                                 full_space=full_space,
@@ -352,15 +352,15 @@ class TestSamplerSelection:
                     ("Accuracy", SimpleNamespace(average_score=0.8)),
                 ])
 
-        with patch("nat.optimizer.parameter_optimizer.optuna.create_study",
+        with patch("nat.config_optimizer.parameters.optimizer.optuna.create_study",
                    side_effect=capture_sampler), \
-             patch("nat.optimizer.parameter_optimizer.load_evaluation_run",
+             patch("nat.config_optimizer.parameters.optimizer.load_evaluation_run",
                    return_value=_DummyEvalRun), \
-             patch("nat.optimizer.parameter_optimizer.apply_suggestions",
+             patch("nat.config_optimizer.parameters.optimizer.apply_suggestions",
                    return_value=base_cfg), \
-             patch("nat.optimizer.parameter_optimizer.pick_trial",
+             patch("nat.config_optimizer.parameters.optimizer.pick_trial",
                    return_value=SimpleNamespace(params={})), \
-             patch("nat.optimizer.pareto_visualizer.create_pareto_visualization"):
+             patch("nat.config_optimizer.parameters.pareto_visualizer.create_pareto_visualization"):
 
             optimize_parameters(base_cfg=base_cfg,
                                 full_space=full_space,
@@ -410,15 +410,15 @@ class TestSamplerSelection:
                     ("Latency", SimpleNamespace(average_score=0.5)),
                 ])
 
-        with patch("nat.optimizer.parameter_optimizer.optuna.create_study",
+        with patch("nat.config_optimizer.parameters.optimizer.optuna.create_study",
                    side_effect=capture_sampler), \
-             patch("nat.optimizer.parameter_optimizer.load_evaluation_run",
+             patch("nat.config_optimizer.parameters.optimizer.load_evaluation_run",
                    return_value=_DummyEvalRun), \
-             patch("nat.optimizer.parameter_optimizer.apply_suggestions",
+             patch("nat.config_optimizer.parameters.optimizer.apply_suggestions",
                    return_value=base_cfg), \
-             patch("nat.optimizer.parameter_optimizer.pick_trial",
+             patch("nat.config_optimizer.parameters.optimizer.pick_trial",
                    return_value=SimpleNamespace(params={})), \
-             patch("nat.optimizer.pareto_visualizer.create_pareto_visualization"):
+             patch("nat.config_optimizer.parameters.pareto_visualizer.create_pareto_visualization"):
 
             optimize_parameters(base_cfg=base_cfg,
                                 full_space=full_space,
@@ -466,17 +466,17 @@ class TestSamplerSelection:
                     ("Latency", SimpleNamespace(average_score=0.5)),
                 ])
 
-        with patch("nat.optimizer.parameter_optimizer.optuna.samplers.GridSampler",
+        with patch("nat.config_optimizer.parameters.optimizer.optuna.samplers.GridSampler",
                    side_effect=capture_grid_sampler), \
-             patch("nat.optimizer.parameter_optimizer.optuna.create_study",
+             patch("nat.config_optimizer.parameters.optimizer.optuna.create_study",
                    return_value=_FakeStudy(["maximize", "minimize"])), \
-             patch("nat.optimizer.parameter_optimizer.load_evaluation_run",
+             patch("nat.config_optimizer.parameters.optimizer.load_evaluation_run",
                    return_value=_DummyEvalRun), \
-             patch("nat.optimizer.parameter_optimizer.apply_suggestions",
+             patch("nat.config_optimizer.parameters.optimizer.apply_suggestions",
                    return_value=base_cfg), \
-             patch("nat.optimizer.parameter_optimizer.pick_trial",
+             patch("nat.config_optimizer.parameters.optimizer.pick_trial",
                    return_value=SimpleNamespace(params={})), \
-             patch("nat.optimizer.pareto_visualizer.create_pareto_visualization"):
+             patch("nat.config_optimizer.parameters.pareto_visualizer.create_pareto_visualization"):
 
             optimize_parameters(base_cfg=base_cfg,
                                 full_space=full_space,
@@ -520,17 +520,17 @@ class TestSamplerSelection:
                     ("Latency", SimpleNamespace(average_score=0.5)),
                 ])
 
-        with patch("nat.optimizer.parameter_optimizer.optuna.samplers.GridSampler",
+        with patch("nat.config_optimizer.parameters.optimizer.optuna.samplers.GridSampler",
                    side_effect=capture_grid_sampler), \
-             patch("nat.optimizer.parameter_optimizer.optuna.create_study",
+             patch("nat.config_optimizer.parameters.optimizer.optuna.create_study",
                    return_value=_FakeStudy(["maximize", "minimize"])), \
-             patch("nat.optimizer.parameter_optimizer.load_evaluation_run",
+             patch("nat.config_optimizer.parameters.optimizer.load_evaluation_run",
                    return_value=_DummyEvalRun), \
-             patch("nat.optimizer.parameter_optimizer.apply_suggestions",
+             patch("nat.config_optimizer.parameters.optimizer.apply_suggestions",
                    return_value=base_cfg), \
-             patch("nat.optimizer.parameter_optimizer.pick_trial",
+             patch("nat.config_optimizer.parameters.optimizer.pick_trial",
                    return_value=SimpleNamespace(params={})), \
-             patch("nat.optimizer.pareto_visualizer.create_pareto_visualization"):
+             patch("nat.config_optimizer.parameters.pareto_visualizer.create_pareto_visualization"):
 
             optimize_parameters(base_cfg=base_cfg,
                                 full_space=full_space,
@@ -567,15 +567,15 @@ class TestSamplerSelection:
                     ("Latency", SimpleNamespace(average_score=0.5)),
                 ])
 
-        with patch("nat.optimizer.parameter_optimizer.optuna.create_study",
+        with patch("nat.config_optimizer.parameters.optimizer.optuna.create_study",
                    side_effect=capture_sampler), \
-             patch("nat.optimizer.parameter_optimizer.load_evaluation_run",
+             patch("nat.config_optimizer.parameters.optimizer.load_evaluation_run",
                    return_value=_DummyEvalRun), \
-             patch("nat.optimizer.parameter_optimizer.apply_suggestions",
+             patch("nat.config_optimizer.parameters.optimizer.apply_suggestions",
                    return_value=base_cfg), \
-             patch("nat.optimizer.parameter_optimizer.pick_trial",
+             patch("nat.config_optimizer.parameters.optimizer.pick_trial",
                    return_value=SimpleNamespace(params={})), \
-             patch("nat.optimizer.pareto_visualizer.create_pareto_visualization"):
+             patch("nat.config_optimizer.parameters.pareto_visualizer.create_pareto_visualization"):
 
             optimize_parameters(base_cfg=base_cfg,
                                 full_space=full_space,
@@ -620,15 +620,15 @@ class TestSamplerSelection:
                     ("Accuracy", SimpleNamespace(average_score=0.8)),
                 ])
 
-        with patch("nat.optimizer.parameter_optimizer.optuna.create_study",
+        with patch("nat.config_optimizer.parameters.optimizer.optuna.create_study",
                    side_effect=capture_sampler), \
-             patch("nat.optimizer.parameter_optimizer.load_evaluation_run",
+             patch("nat.config_optimizer.parameters.optimizer.load_evaluation_run",
                    return_value=_DummyEvalRun), \
-             patch("nat.optimizer.parameter_optimizer.apply_suggestions",
+             patch("nat.config_optimizer.parameters.optimizer.apply_suggestions",
                    return_value=base_cfg), \
-             patch("nat.optimizer.parameter_optimizer.pick_trial",
+             patch("nat.config_optimizer.parameters.optimizer.pick_trial",
                    return_value=SimpleNamespace(params={})), \
-             patch("nat.optimizer.pareto_visualizer.create_pareto_visualization"):
+             patch("nat.config_optimizer.parameters.pareto_visualizer.create_pareto_visualization"):
 
             optimize_parameters(base_cfg=base_cfg,
                                 full_space=full_space,
@@ -676,15 +676,15 @@ class TestSamplerSelection:
                     ("Latency", SimpleNamespace(average_score=0.5)),
                 ])
 
-        with patch("nat.optimizer.parameter_optimizer.optuna.create_study",
+        with patch("nat.config_optimizer.parameters.optimizer.optuna.create_study",
                    side_effect=capture_sampler), \
-             patch("nat.optimizer.parameter_optimizer.load_evaluation_run",
+             patch("nat.config_optimizer.parameters.optimizer.load_evaluation_run",
                    return_value=_DummyEvalRun), \
-             patch("nat.optimizer.parameter_optimizer.apply_suggestions",
+             patch("nat.config_optimizer.parameters.optimizer.apply_suggestions",
                    return_value=base_cfg), \
-             patch("nat.optimizer.parameter_optimizer.pick_trial",
+             patch("nat.config_optimizer.parameters.optimizer.pick_trial",
                    return_value=SimpleNamespace(params={})), \
-             patch("nat.optimizer.pareto_visualizer.create_pareto_visualization"):
+             patch("nat.config_optimizer.parameters.pareto_visualizer.create_pareto_visualization"):
 
             optimize_parameters(base_cfg=base_cfg,
                                 full_space=full_space,
@@ -725,15 +725,15 @@ class TestSamplerSelection:
                     ("Latency", SimpleNamespace(average_score=0.5)),
                 ])
 
-        with patch("nat.optimizer.parameter_optimizer.optuna.create_study",
+        with patch("nat.config_optimizer.parameters.optimizer.optuna.create_study",
                    side_effect=capture_sampler), \
-             patch("nat.optimizer.parameter_optimizer.load_evaluation_run",
+             patch("nat.config_optimizer.parameters.optimizer.load_evaluation_run",
                    return_value=_DummyEvalRun), \
-             patch("nat.optimizer.parameter_optimizer.apply_suggestions",
+             patch("nat.config_optimizer.parameters.optimizer.apply_suggestions",
                    return_value=base_cfg), \
-             patch("nat.optimizer.parameter_optimizer.pick_trial",
+             patch("nat.config_optimizer.parameters.optimizer.pick_trial",
                    return_value=SimpleNamespace(params={})), \
-             patch("nat.optimizer.pareto_visualizer.create_pareto_visualization"):
+             patch("nat.config_optimizer.parameters.pareto_visualizer.create_pareto_visualization"):
 
             optimize_parameters(base_cfg=base_cfg,
                                 full_space=full_space,
@@ -783,17 +783,17 @@ class TestGridSearchIntegration:
                     ("Latency", SimpleNamespace(average_score=0.5)),
                 ])
 
-        with patch("nat.optimizer.parameter_optimizer.optuna.samplers.GridSampler",
+        with patch("nat.config_optimizer.parameters.optimizer.optuna.samplers.GridSampler",
                    side_effect=capture_grid_sampler), \
-             patch("nat.optimizer.parameter_optimizer.optuna.create_study",
+             patch("nat.config_optimizer.parameters.optimizer.optuna.create_study",
                    return_value=_FakeStudy(["maximize", "minimize"])), \
-             patch("nat.optimizer.parameter_optimizer.load_evaluation_run",
+             patch("nat.config_optimizer.parameters.optimizer.load_evaluation_run",
                    return_value=_DummyEvalRun), \
-             patch("nat.optimizer.parameter_optimizer.apply_suggestions",
+             patch("nat.config_optimizer.parameters.optimizer.apply_suggestions",
                    return_value=base_cfg), \
-             patch("nat.optimizer.parameter_optimizer.pick_trial",
+             patch("nat.config_optimizer.parameters.optimizer.pick_trial",
                    return_value=SimpleNamespace(params={})), \
-             patch("nat.optimizer.pareto_visualizer.create_pareto_visualization"):
+             patch("nat.config_optimizer.parameters.pareto_visualizer.create_pareto_visualization"):
 
             optimize_parameters(base_cfg=base_cfg,
                                 full_space=full_space,
@@ -842,17 +842,17 @@ class TestGridSearchIntegration:
                     ("Latency", SimpleNamespace(average_score=0.5)),
                 ])
 
-        with patch("nat.optimizer.parameter_optimizer.optuna.samplers.GridSampler",
+        with patch("nat.config_optimizer.parameters.optimizer.optuna.samplers.GridSampler",
                    side_effect=capture_grid_sampler), \
-             patch("nat.optimizer.parameter_optimizer.optuna.create_study",
+             patch("nat.config_optimizer.parameters.optimizer.optuna.create_study",
                    return_value=_FakeStudy(["maximize", "minimize"])), \
-             patch("nat.optimizer.parameter_optimizer.load_evaluation_run",
+             patch("nat.config_optimizer.parameters.optimizer.load_evaluation_run",
                    return_value=_DummyEvalRun), \
-             patch("nat.optimizer.parameter_optimizer.apply_suggestions",
+             patch("nat.config_optimizer.parameters.optimizer.apply_suggestions",
                    return_value=base_cfg), \
-             patch("nat.optimizer.parameter_optimizer.pick_trial",
+             patch("nat.config_optimizer.parameters.optimizer.pick_trial",
                    return_value=SimpleNamespace(params={})), \
-             patch("nat.optimizer.pareto_visualizer.create_pareto_visualization"):
+             patch("nat.config_optimizer.parameters.pareto_visualizer.create_pareto_visualization"):
 
             optimize_parameters(base_cfg=base_cfg,
                                 full_space=full_space,
@@ -900,17 +900,17 @@ class TestGridSearchIntegration:
                     ("Latency", SimpleNamespace(average_score=0.5)),
                 ])
 
-        with patch("nat.optimizer.parameter_optimizer.optuna.samplers.GridSampler",
+        with patch("nat.config_optimizer.parameters.optimizer.optuna.samplers.GridSampler",
                    side_effect=capture_grid_sampler), \
-             patch("nat.optimizer.parameter_optimizer.optuna.create_study",
+             patch("nat.config_optimizer.parameters.optimizer.optuna.create_study",
                    return_value=_FakeStudy(["maximize"])), \
-             patch("nat.optimizer.parameter_optimizer.load_evaluation_run",
+             patch("nat.config_optimizer.parameters.optimizer.load_evaluation_run",
                    return_value=_DummyEvalRun), \
-             patch("nat.optimizer.parameter_optimizer.apply_suggestions",
+             patch("nat.config_optimizer.parameters.optimizer.apply_suggestions",
                    return_value=base_cfg), \
-             patch("nat.optimizer.parameter_optimizer.pick_trial",
+             patch("nat.config_optimizer.parameters.optimizer.pick_trial",
                    return_value=SimpleNamespace(params={})), \
-             patch("nat.optimizer.pareto_visualizer.create_pareto_visualization"):
+             patch("nat.config_optimizer.parameters.pareto_visualizer.create_pareto_visualization"):
 
             optimize_parameters(base_cfg=base_cfg,
                                 full_space=full_space,
@@ -954,17 +954,17 @@ class TestGridSearchIntegration:
                     ("Latency", SimpleNamespace(average_score=0.5)),
                 ])
 
-        with patch("nat.optimizer.parameter_optimizer.optuna.samplers.GridSampler",
+        with patch("nat.config_optimizer.parameters.optimizer.optuna.samplers.GridSampler",
                    side_effect=capture_grid_sampler), \
-             patch("nat.optimizer.parameter_optimizer.optuna.create_study",
+             patch("nat.config_optimizer.parameters.optimizer.optuna.create_study",
                    return_value=_FakeStudy(["minimize"])), \
-             patch("nat.optimizer.parameter_optimizer.load_evaluation_run",
+             patch("nat.config_optimizer.parameters.optimizer.load_evaluation_run",
                    return_value=_DummyEvalRun), \
-             patch("nat.optimizer.parameter_optimizer.apply_suggestions",
+             patch("nat.config_optimizer.parameters.optimizer.apply_suggestions",
                    return_value=base_cfg), \
-             patch("nat.optimizer.parameter_optimizer.pick_trial",
+             patch("nat.config_optimizer.parameters.optimizer.pick_trial",
                    return_value=SimpleNamespace(params={})), \
-             patch("nat.optimizer.pareto_visualizer.create_pareto_visualization"):
+             patch("nat.config_optimizer.parameters.pareto_visualizer.create_pareto_visualization"):
 
             optimize_parameters(base_cfg=base_cfg,
                                 full_space=full_space,
@@ -1012,17 +1012,17 @@ class TestGridSearchIntegration:
                     ("Latency", SimpleNamespace(average_score=0.5)),
                 ])
 
-        with patch("nat.optimizer.parameter_optimizer.optuna.samplers.GridSampler",
+        with patch("nat.config_optimizer.parameters.optimizer.optuna.samplers.GridSampler",
                    side_effect=capture_grid_sampler), \
-             patch("nat.optimizer.parameter_optimizer.optuna.create_study",
+             patch("nat.config_optimizer.parameters.optimizer.optuna.create_study",
                    return_value=_FakeStudy(["maximize", "minimize"])), \
-             patch("nat.optimizer.parameter_optimizer.load_evaluation_run",
+             patch("nat.config_optimizer.parameters.optimizer.load_evaluation_run",
                    return_value=_DummyEvalRun), \
-             patch("nat.optimizer.parameter_optimizer.apply_suggestions",
+             patch("nat.config_optimizer.parameters.optimizer.apply_suggestions",
                    return_value=base_cfg), \
-             patch("nat.optimizer.parameter_optimizer.pick_trial",
+             patch("nat.config_optimizer.parameters.optimizer.pick_trial",
                    return_value=SimpleNamespace(params={})), \
-             patch("nat.optimizer.pareto_visualizer.create_pareto_visualization"):
+             patch("nat.config_optimizer.parameters.pareto_visualizer.create_pareto_visualization"):
 
             optimize_parameters(base_cfg=base_cfg,
                                 full_space=full_space,
@@ -1071,17 +1071,17 @@ class TestGridSearchIntegration:
                     ("Latency", SimpleNamespace(average_score=0.5)),
                 ])
 
-        with patch("nat.optimizer.parameter_optimizer.optuna.samplers.GridSampler",
+        with patch("nat.config_optimizer.parameters.optimizer.optuna.samplers.GridSampler",
                    side_effect=capture_grid_sampler), \
-             patch("nat.optimizer.parameter_optimizer.optuna.create_study",
+             patch("nat.config_optimizer.parameters.optimizer.optuna.create_study",
                    return_value=_FakeStudy(["maximize"])), \
-             patch("nat.optimizer.parameter_optimizer.load_evaluation_run",
+             patch("nat.config_optimizer.parameters.optimizer.load_evaluation_run",
                    return_value=_DummyEvalRun), \
-             patch("nat.optimizer.parameter_optimizer.apply_suggestions",
+             patch("nat.config_optimizer.parameters.optimizer.apply_suggestions",
                    return_value=base_cfg), \
-             patch("nat.optimizer.parameter_optimizer.pick_trial",
+             patch("nat.config_optimizer.parameters.optimizer.pick_trial",
                    return_value=SimpleNamespace(params={})), \
-             patch("nat.optimizer.pareto_visualizer.create_pareto_visualization"):
+             patch("nat.config_optimizer.parameters.pareto_visualizer.create_pareto_visualization"):
 
             optimize_parameters(base_cfg=base_cfg,
                                 full_space=full_space,
@@ -1108,10 +1108,10 @@ class TestGridSearchIntegration:
         run_cfg = _make_run_config(base_cfg)
 
         with pytest.raises(ValueError, match="requires 'step' to be specified"):
-            with patch("nat.optimizer.parameter_optimizer.load_evaluation_run"), \
-                 patch("nat.optimizer.parameter_optimizer.apply_suggestions",
+            with patch("nat.config_optimizer.parameters.optimizer.load_evaluation_run"), \
+                 patch("nat.config_optimizer.parameters.optimizer.apply_suggestions",
                        return_value=base_cfg), \
-                 patch("nat.optimizer.pareto_visualizer.create_pareto_visualization"):
+                 patch("nat.config_optimizer.parameters.pareto_visualizer.create_pareto_visualization"):
 
                 optimize_parameters(base_cfg=base_cfg,
                                     full_space=full_space,
