+++ packages/nvidia_nat_config_optimizer/tests/test_update_helpers.py
@@ -15,8 +15,8 @@
 
 from pydantic import BaseModel
 
-from nat.optimizer.update_helpers import apply_suggestions
-from nat.optimizer.update_helpers import nest_updates
+from nat.config_optimizer.update_helpers import apply_suggestions
+from nat.config_optimizer.update_helpers import nest_updates
 
 
 def test_nest_updates_merges_nested_keys():
