+++ packages/nvidia_nat_core/src/nat/cli/commands/optimize.py
@@ -22,7 +22,7 @@ import click
 from nat.data_models.optimizer import OptimizerRunConfig
 
 try:
-    from nat.optimizer.optimizer_runtime import optimize_config
+    from nat.config_optimizer.optimizer_runtime import optimize_config
 except ImportError as _e:
     optimize_config = None
     _optimizer_import_error = _e
@@ -71,7 +71,7 @@ def optimizer_command(ctx, **kwargs) -> None:
 
 async def run_optimizer(config: OptimizerRunConfig):
     if optimize_config is None:
-        raise click.UsageError("The optimizer is not installed. Install it with: pip install nvidia-nat-optimizer\n"
+        raise click.UsageError("The optimizer is not installed. Install it with: pip install nvidia-nat-config-optimizer\n"
                                "Or install the full toolkit: pip install nvidia-nat") from _optimizer_import_error
     await optimize_config(config)
