+++ packages/nvidia_nat_core/tests/nat/data_models/test_optimizer_oracle_feedback.py
@@ -3,7 +3,7 @@
 # SPDX-License-Identifier: Apache-2.0
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use it except in compliance with the License.
+# you may not use this file except in compliance with the License.
 # You may obtain a copy of the License at
 #
 # http://www.apache.org/licenses/LICENSE-2.0
@@ -15,33 +15,36 @@
 # limitations under the License.
 
 """
-Tests for PromptGAOptimizationConfig and extra (oracle) fields.
+Tests for GAPromptOptimizationConfig oracle feedback fields.
 
-Oracle feedback is implementation-specific; it is not part of the shared
-PromptGAOptimizationConfig contract. Extra fields (e.g. oracle_feedback_mode)
-are allowed and stored in model_extra so YAML can still specify them.
-Validation of oracle_* is done by GAOptimizerConfig in nvidia_nat_optimizer.
+Oracle feedback fields are typed directly on GAPromptOptimizationConfig
+with proper defaults and validation.
 """
 
-from nat.data_models.optimizer import PromptGAOptimizationConfig
+from nat.data_models.optimizer import GAPromptOptimizationConfig
 
 
-class TestPromptGAOptimizationConfigExtra:
-    """PromptGAOptimizationConfig accepts oracle_* as extra and stores in model_extra."""
+class TestGAPromptOptimizationConfigOracleFeedback:
+    """GAPromptOptimizationConfig has typed oracle feedback fields."""
 
-    def test_accepts_oracle_extra_fields(self):
-        """Oracle feedback keys are accepted as extra and stored in model_extra."""
-        config = PromptGAOptimizationConfig(
+    def test_oracle_feedback_fields_as_typed_attributes(self):
+        """Oracle feedback fields are proper typed attributes with defaults."""
+        config = GAPromptOptimizationConfig(
             oracle_feedback_mode="always",
             oracle_feedback_worst_n=3,
             oracle_feedback_max_chars=2000,
         )
-        assert config.model_extra is not None
-        assert config.model_extra.get("oracle_feedback_mode") == "always"
-        assert config.model_extra.get("oracle_feedback_worst_n") == 3
-        assert config.model_extra.get("oracle_feedback_max_chars") == 2000
+        assert config.oracle_feedback_mode == "always"
+        assert config.oracle_feedback_worst_n == 3
+        assert config.oracle_feedback_max_chars == 2000
 
-    def test_no_oracle_extra_by_default(self):
-        """Without oracle keys, model_extra is empty or absent."""
-        config = PromptGAOptimizationConfig()
-        assert getattr(config, "model_extra", None) is None or config.model_extra == {}
+    def test_oracle_feedback_defaults(self):
+        """Without oracle keys, defaults are applied."""
+        config = GAPromptOptimizationConfig()
+        assert config.oracle_feedback_mode == "never"
+        assert config.oracle_feedback_worst_n == 5
+        assert config.oracle_feedback_max_chars == 4000
+        assert config.oracle_feedback_fitness_threshold == 0.3
+        assert config.oracle_feedback_stagnation_generations == 3
+        assert config.oracle_feedback_fitness_variance_threshold == 0.01
+        assert config.oracle_feedback_diversity_threshold == 0.5
